Separations

+++ Backend(Server)/Controler/AdminControllers/Courses.js
@@ -1,7 +1,8 @@
-var coursedoc = require("../Models/Course");
+var coursedoc = require("../../Models/Course");
 
-module.exports.Add = async (req, res, next) => {
+module.exports.Add = async (req, res) => {
   try {
+    if(!req.user)return await res.json("Timed Out")
     const course = await coursedoc.create(req.body);
     console.log("course added", course);
     await res.json(course);
@@ -9,8 +10,9 @@ module.exports.Add = async (req, res, next) => {
     console.log(err);
   }
 };
-module.exports.Showall = async (req, res, next) => {
+module.exports.Showall = async (req, res) => {
   try {
+    if(!req.user)return await res.json("Timed Out")
     const course = await coursedoc.find({});
     console.log("all courses", course);
     await res.json(course);
@@ -19,8 +21,9 @@ module.exports.Showall = async (req, res, next) => {
   }
 };
 
-module.exports.ShowOne = async (req, res, next) => {
+module.exports.ShowOne = async (req, res) => {
   try {
+    if(!req.user)return await res.json("Timed Out")
     const course = await coursedoc.findById(req.params.id);
     res.json(course);
   } catch (err) {
@@ -28,8 +31,9 @@ module.exports.ShowOne = async (req, res, next) => {
   }
 };
 
-module.exports.Delete = async (req, res, next) => {
+module.exports.Delete = async (req, res) => {
   try {
+    if(!req.user)return await res.json("Timed Out")
     const course = await coursedoc.deleteOne({ _id: req.params.id });
     console.log("all courses", course);
     await res.json(course);
@@ -37,8 +41,9 @@ module.exports.Delete = async (req, res, next) => {
     console.log(err);
   }
 };
-module.exports.Update = async (req, res, next) => {
+module.exports.Update = async (req, res) => {
   try {
+    if(!req.user)return await res.json("Timed Out")
     const course = await coursedoc.findOneAndUpdate({ _id: req.params.id },req.body);
     console.log("all courses", course);
     await res.json(course);
