+++ Backend(Server)/Controler/AdminControllers/Faculty.js
@@ -1,20 +1,24 @@
-var coursedoc = require("../Models/Course");
-var facultydoc = require("../Models/Faculty");
+var coursedoc = require("../../Models/Course");
+var facultydoc = require("../../Models/Faculty");
+var bcrypt = require("bcrypt");
 
-module.exports.Add = async (req, res, next) => {
+module.exports.Add = async (req, res) => {
   try {
+    if(!req.user)return await res.json("Timed Out")
     const {FirstName,SecondName,Email,Password,Allocated,Degree}=req.body
     const cor = await coursedoc.findOne({Allocated})
     if (!cor) return res.status('400').json('no course')
-    const faculty = await facultydoc.create({FirstName,SecondName,Email,Password,Course:cor._id});
+    newPassword = await bcrypt.hash(Password, 12);
+    const faculty = await facultydoc.create({FirstName,SecondName,Email,Password:newPassword,Course:cor._id});
     console.log("faculty added", faculty);
     await res.json(faculty);
   } catch (err) {
     console.log(err);
   }
 }
-module.exports.Showall = async (req, res, next) => {
+module.exports.Showall = async (req, res) => {
   try {
+    if(!req.user)return await res.json("Timed Out")
     console.log('insiede')
     const faculty = await facultydoc.find({}).populate("Course");
     console.log("all faculty", faculty);
@@ -24,8 +28,9 @@ module.exports.Showall = async (req, res, next) => {
   }
 }
 
-module.exports.Delete = async (req, res, next) => {
+module.exports.Delete = async (req, res) => {
   try {
+    if(!req.user)return await res.json("Timed Out")
     const faculty = await facultydoc.deleteOne({ _id: req.params.id });
     console.log("all faulty", faculty);
     await res.json(faculty);
@@ -35,7 +40,7 @@ module.exports.Delete = async (req, res, next) => {
 };
 
 /*
-module.exports.ShowOne = async (req, res, next) => {
+module.exports.ShowOne = async (req, res) => {
   try {
     const course = await coursedoc.findById(req.params.id);
     res.json(course);
@@ -43,8 +48,9 @@ module.exports.ShowOne = async (req, res, next) => {
     console.log(err);
   }
 };
-module.exports.Update = async (req, res, next) => {
+module.exports.Update = async (req, res) => {
   try {
+    if(!req.user.isAdmin||!req.user)return await res.json("Timed Out")
     const course = await coursedoc.findOneAndUpdate({ _id: req.params.id },req.body);
     console.log("all courses", course);
     await res.json(course);
